Test writing documentation with markdown and html
Markdown likely won't be used since it required inline html to set the background colors of table cells. 

Likely will use html. Potentially paired with a static site generator to make things easier. If that path is taken the static site could be generated by github actions and hosted as a github site.