Add ports to VM
First pass on ports. Ports are memory address that the VM can write to in order to control and communicate with robot hardware. The first 256 bytes of the VM are reserved for ports and other VM data. At the moment there are only 3 ports, with only P_SPEDOMETER being implemented. I'm not completely happy with the implementation so far. I'll do a cleanup pass after implementing some more ports and hardware since it'll be more clear what they require.