Robot program live reload
Added live reload setting that's enabled by default. If a robots source code file is edited and saved it automatically gets recompiled and reloaded. This also resets the robots VM (registers, memory, stack, variables) to the default state.This will save a lot of development & debugging time but also make it much easier to design and test robots. If the there's an error or invalid syntax on reload it'll keep using the existing program. Some things that need to be worked on:
- Error messages aren't very helpful. They should specify the line number and catch simple errors like using invalid registers or instruction/register typos.
- There are some compilation bugs. I had one case where the whole app froze on reload. At some point I'll intentionally throw a bunch of invalid code at it and fix any crashes or freezing I find.
- It'd be nice if error messages were shown in the UI instead of the external console.