Make robots framerate and cycle rate independent
Robot behavior should now be the same regardless of framerate and cycle rate. The only exception is that robot movement is choppy at low cycle rates. This is expected since it's being updated less frequently.

This commit also fixes a crash in the bullet deletion code and removes the old robot update loop which was accidentally left in.