Add ipo/opo instructions for reading/writing ports
This is being used instead of load/store to be similar to the original and also so some interacting with some ports can be faster or slower than reading/writing memory.

One instruction this added, OpoVal uses slightly different encoding than the rest to fit the opcode, port, and value in a 32bit instruction. This normally wouldn't be possible since the port and value are 16bits. I made it work by using less bits for the port, which is fine since there are fewer than 100 ports.

The problem with this is that a few special cases were needed for the compiler to properly handle it. I'm going to think about this a bit and fix it in the next few days.