Fix File::ReadAll()
Just using std::stringstream instead of trying to read it char by char. The old version had a bug where the last character in the file would be repeated many times if the file doesn't end with a newline. This version doesn't have that issue and is easier to understand.